Split up main checkout tests file
