+++ client/my-sites/checkout/composite-checkout/test/composite-checkout.js
@@ -3,15 +3,7 @@
  */
 import { StripeHookProvider } from '@automattic/calypso-stripe';
 import { ShoppingCartProvider, createShoppingCartManagerClient } from '@automattic/shopping-cart';
-import {
-	render,
-	fireEvent,
-	screen,
-	within,
-	waitFor,
-	act,
-	waitForElementToBeRemoved,
-} from '@testing-library/react';
+import { render, fireEvent, screen, within, waitFor, act } from '@testing-library/react';
 import nock from 'nock';
 import { Provider as ReduxProvider } from 'react-redux';
 import { navigate } from 'calypso/lib/navigate';
@@ -25,8 +17,6 @@ import CompositeCheckout from '../composite-checkout';
 import {
 	siteId,
 	domainProduct,
-	domainTransferProduct,
-	planWithBundledDomain,
 	planWithoutDomain,
 	fetchStripeConfiguration,
 	mockSetCartEndpointWith,
@@ -34,8 +24,6 @@ import {
 	getActivePersonalPlanDataForType,
 	createTestReduxStore,
 	countryList,
-	gSuiteProduct,
-	caDomainProduct,
 } from './util';
 
 /* eslint-disable jest/no-conditional-expect */
@@ -171,401 +159,6 @@ describe( 'CompositeCheckout', () => {
 			screen
 				.getAllByLabelText( 'Total' )
 				.map( ( element ) => expect( element ).toHaveTextContent( 'R$156' ) );
-		} );
-	} );
-
-	it( 'renders the paypal payment method option', async () => {
-		render( <MyCheckout />, container );
-		await waitFor( () => {
-			expect( screen.getByText( 'PayPal' ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'does not render the full credits payment method option when no credits are available', async () => {
-		render( <MyCheckout />, container );
-		await waitFor( () => {
-			expect( screen.queryByText( /WordPress.com Credits:/ ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'does not render the full credits payment method option when partial credits are available', async () => {
-		const cartChanges = { credits_integer: 15400, credits_display: 'R$154' };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.queryByText( /WordPress.com Credits:/ ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders the paypal payment method option when partial credits are available', async () => {
-		const cartChanges = { credits_integer: 15400, credits_display: 'R$154' };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.getByText( 'PayPal' ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders the full credits payment method option when full credits are available', async () => {
-		const cartChanges = {
-			sub_total_integer: 0,
-			sub_total_display: '0',
-			credits_integer: 15600,
-			credits_display: 'R$156',
-		};
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.getByText( /WordPress.com Credits:/ ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'does not render the other payment method options when full credits are available', async () => {
-		const cartChanges = {
-			sub_total_integer: 0,
-			sub_total_display: '0',
-			credits_integer: 15600,
-			credits_display: 'R$156',
-		};
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.queryByText( 'PayPal' ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'does not render the free payment method option when the purchase is not free', async () => {
-		render( <MyCheckout />, container );
-		await waitFor( () => {
-			expect( screen.queryByText( 'Free Purchase' ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'does not render the paypal payment method option when the purchase is free', async () => {
-		const cartChanges = { total_cost_integer: 0, total_cost_display: '0' };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.queryByText( 'PayPal' ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'does not render the full credits payment method option when full credits are available but the purchase is free', async () => {
-		const cartChanges = {
-			sub_total_integer: 0,
-			sub_total_display: '0',
-			total_tax_integer: 0,
-			total_tax_display: 'R$0',
-			total_cost_integer: 0,
-			total_cost_display: '0',
-			credits_integer: 15600,
-			credits_display: 'R$156',
-		};
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.queryByText( /WordPress.com Credits:/ ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders the free payment method option when the purchase is free', async () => {
-		const cartChanges = { total_cost_integer: 0, total_cost_display: '0' };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.getByText( 'Free Purchase' ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'does not render the contact step when the purchase is free', async () => {
-		const cartChanges = { total_cost_integer: 0, total_cost_display: '0' };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect(
-				screen.queryByText( /Enter your (billing|contact) information/ )
-			).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders the contact step when the purchase is not free', async () => {
-		render( <MyCheckout />, container );
-		await waitFor( () => {
-			expect( screen.getByText( /Enter your (billing|contact) information/ ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders the tax fields only when no domain is in the cart', async () => {
-		const cartChanges = { products: [ planWithoutDomain ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-			expect( screen.queryByText( 'Phone' ) ).not.toBeInTheDocument();
-			expect( screen.queryByText( 'Email' ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders the domain fields when a domain is in the cart', async () => {
-		const cartChanges = { products: [ planWithBundledDomain, domainProduct ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Phone' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Email' ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders the domain fields when a domain transfer is in the cart', async () => {
-		const cartChanges = { products: [ planWithBundledDomain, domainTransferProduct ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Phone' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Email' ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'does not render country-specific domain fields when no country has been chosen and a domain is in the cart', async () => {
-		const cartChanges = { products: [ planWithBundledDomain, domainProduct ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Phone' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Email' ) ).toBeInTheDocument();
-			expect( screen.queryByText( 'Address' ) ).not.toBeInTheDocument();
-			expect( screen.queryByText( 'City' ) ).not.toBeInTheDocument();
-			expect( screen.queryByText( 'State' ) ).not.toBeInTheDocument();
-			expect( screen.queryByText( 'ZIP code' ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders country-specific domain fields when a country has been chosen and a domain is in the cart', async () => {
-		const cartChanges = { products: [ planWithBundledDomain, domainProduct ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			fireEvent.change( screen.getByLabelText( 'Country' ), { target: { value: 'US' } } );
-		} );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Phone' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Email' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Address' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'City' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'State' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'ZIP code' ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders domain fields with postal code when a country with postal code support has been chosen and a plan is in the cart', async () => {
-		const cartChanges = { products: [ planWithoutDomain ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			fireEvent.change( screen.getByLabelText( 'Country' ), { target: { value: 'US' } } );
-		} );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Postal code' ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders domain fields except postal code when a country without postal code support has been chosen and a plan is in the cart', async () => {
-		const cartChanges = { products: [ planWithoutDomain ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			fireEvent.change( screen.getByLabelText( 'Country' ), { target: { value: 'CW' } } );
-		} );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-			expect( screen.queryByText( 'Postal code' ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders domain fields with postal code when a country with postal code support has been chosen and a domain is in the cart', async () => {
-		const cartChanges = { products: [ planWithBundledDomain, domainProduct ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			fireEvent.change( screen.getByLabelText( 'Country' ), { target: { value: 'US' } } );
-		} );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Phone' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Email' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'ZIP code' ) ).toBeInTheDocument();
-		} );
-	} );
-
-	it( 'renders domain fields except postal code when a country without postal code support has been chosen and a domain is in the cart', async () => {
-		const cartChanges = { products: [ planWithBundledDomain, domainProduct ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			fireEvent.change( screen.getByLabelText( 'Country' ), { target: { value: 'CW' } } );
-		} );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Phone' ) ).toBeInTheDocument();
-			expect( screen.getByText( 'Email' ) ).toBeInTheDocument();
-			expect( screen.queryByText( 'Postal Code' ) ).not.toBeInTheDocument();
-			expect( screen.queryByText( 'Postal code' ) ).not.toBeInTheDocument();
-			expect( screen.queryByText( 'ZIP code' ) ).not.toBeInTheDocument();
-		} );
-	} );
-
-	it( 'does not complete the contact step when the contact step button has not been clicked', async () => {
-		const cartChanges = { products: [ planWithoutDomain ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		await waitFor( () => {
-			expect( screen.getByText( 'Country' ) ).toBeInTheDocument();
-		} );
-		expect( screen.queryByTestId( 'payment-method-step--visible' ) ).not.toBeInTheDocument();
-	} );
-
-	it.each( [
-		{ complete: 'does', valid: 'valid', name: 'plan', email: 'fails', logged: 'in' },
-		{ complete: 'does not', valid: 'invalid', name: 'plan', email: 'fails', logged: 'in' },
-		{ complete: 'does', valid: 'valid', name: 'domain', email: 'fails', logged: 'in' },
-		{ complete: 'does not', valid: 'invalid', name: 'domain', email: 'fails', logged: 'in' },
-		{ complete: 'does', valid: 'valid', name: 'gsuite', email: 'fails', logged: 'in' },
-		{ complete: 'does not', valid: 'invalid', name: 'gsuite', email: 'fails', logged: 'in' },
-		{ complete: 'does', valid: 'valid', name: 'plan', email: 'passes', logged: 'out' },
-		{ complete: 'does not', valid: 'invalid', name: 'plan', email: 'passes', logged: 'out' },
-		{ complete: 'does not', valid: 'valid', name: 'domain', email: 'fails', logged: 'out' },
-		{ complete: 'does not', valid: 'invalid', name: 'plan', email: 'fails', logged: 'out' },
-	] )(
-		'$complete complete the contact step when validation is $valid with $name in the cart while logged-$logged and signup validation $email',
-		async ( { complete, valid, name, email, logged } ) => {
-			const product = ( () => {
-				switch ( name ) {
-					case 'plan':
-						return planWithoutDomain;
-					case 'domain':
-						return caDomainProduct;
-					case 'gsuite':
-						return gSuiteProduct;
-				}
-			} )();
-			const endpointPath = ( () => {
-				switch ( name ) {
-					case 'plan':
-						return '/rest/v1.1/me/tax-contact-information/validate';
-					case 'domain':
-						return '/rest/v1.2/me/domain-contact-information/validate';
-					case 'gsuite':
-						return '/rest/v1.1/me/google-apps/validate';
-				}
-			} )();
-			const validContactDetails = {
-				postal_code: '10001',
-				country_code: 'US',
-				email: 'test@example.com',
-			};
-			nock.cleanAll();
-			const messages = ( () => {
-				if ( valid === 'valid' ) {
-					return undefined;
-				}
-				if ( name === 'domain' ) {
-					return {
-						postal_code: [ 'Postal code error message' ],
-						'extra.ca.cira_agreement_accepted': [ 'Missing CIRA agreement' ],
-					};
-				}
-				return {
-					postal_code: [ 'Postal code error message' ],
-				};
-			} )();
-			nock( 'https://public-api.wordpress.com' )
-				.post( endpointPath, ( body ) => {
-					if (
-						body.contact_information.postal_code === validContactDetails.postal_code &&
-						body.contact_information.country_code === validContactDetails.country_code
-					) {
-						if ( name === 'domain' ) {
-							return (
-								body.contact_information.email === validContactDetails.email &&
-								body.domain_names[ 0 ] === product.meta
-							);
-						}
-						if ( name === 'gsuite' ) {
-							return body.domain_names[ 0 ] === product.meta;
-						}
-						return true;
-					}
-				} )
-				.reply( 200, {
-					success: valid === 'valid',
-					messages,
-				} );
-			nock( 'https://public-api.wordpress.com' )
-				.post( '/rest/v1.1/signups/validation/user/', ( body ) => {
-					return (
-						body.locale === 'en' &&
-						body.is_from_registrationless_checkout === true &&
-						body.email === validContactDetails.email
-					);
-				} )
-				.reply( 200, () => {
-					if ( logged === 'out' && email === 'fails' ) {
-						return {
-							success: false,
-							messages: { email: { taken: 'An account with this email already exists.' } },
-						};
-					}
-
-					return {
-						success: email === 'passes',
-					};
-				} );
-
-			render(
-				<MyCheckout
-					cartChanges={ { products: [ product ] } }
-					additionalProps={ { isLoggedOutCart: logged === 'out' } }
-				/>,
-				container
-			);
-
-			// Wait for the cart to load
-			await screen.findByText( 'Country' );
-
-			// Fill in the contact form
-			if ( name === 'domain' || logged === 'out' ) {
-				fireEvent.change( screen.getByLabelText( 'Email' ), {
-					target: { value: validContactDetails.email },
-				} );
-			}
-			fireEvent.change( screen.getByLabelText( 'Country' ), {
-				target: { value: validContactDetails.country_code },
-			} );
-			fireEvent.change( screen.getByLabelText( /(Postal|ZIP) code/i ), {
-				target: { value: validContactDetails.postal_code },
-			} );
-			fireEvent.click( await screen.findByText( 'Continue' ) );
-
-			// Wait for the validation to complete
-			await waitForElementToBeRemoved( () => screen.queryAllByText( 'Please wait…' ) );
-
-			if ( complete === 'does' ) {
-				expect( await screen.findByTestId( 'payment-method-step--visible' ) ).toBeInTheDocument();
-			} else {
-				await expect( screen.findByTestId( 'payment-method-step--visible' ) ).toNeverAppear();
-
-				// Make sure the error message is displayed
-				if ( valid !== 'valid' ) {
-					if ( logged === 'out' && email === 'fails' ) {
-						expect(
-							screen.getByText( ( content ) =>
-								content.startsWith( 'That email address is already in use' )
-							)
-						);
-					} else if ( email === 'passes' ) {
-						expect( screen.getByText( 'Postal code error message' ) ).toBeInTheDocument();
-					}
-
-					if ( name === 'domain' ) {
-						expect( screen.getByText( 'Missing CIRA agreement' ) ).toBeInTheDocument();
-					}
-				}
-			}
-		}
-	);
-
-	it( 'renders the checkout summary', async () => {
-		render( <MyCheckout />, container );
-		await waitFor( () => {
-			expect( screen.getByText( 'Purchase Details' ) ).toBeInTheDocument();
-			expect( navigate ).not.toHaveBeenCalled();
 		} );
 	} );
